requirements.txt
streamlit
replicate
python-dotenv